.ipynb_checkpoints/requirements-checkpoint.txt
torch==1.10.1
torchvision==0.11.2
matplotlib==3.4.2
numpy==1.21.0
Pillow>=9.0.0
